Auto search trackers for ID and description
Don't rely only on PTP or require torrent ID, auto search instead. No idea if I broke the manual arguments, will check that later.

HDB with is_disc working for ID.

BLU searching is working, but needs to be updated to search file list in the API instead. Also waiting for a PR to be approved so that we can search folder names.